+++ src/commands/build-react-app.js
@@ -1,5 +1,6 @@
 import path from 'path'
 
+import {getDefaultHTMLConfig} from '../appConfig'
 import {copyPublicDir} from '../utils'
 import webpackBuild from '../webpackBuild'
 
@@ -18,9 +19,7 @@ let buildConfig = () => {
       publicPath: '/'
     },
     plugins: {
-      html: {
-        template: path.resolve('src/index.html')
-      },
+      html: getDefaultHTMLConfig(),
       vendorChunkName: 'vendor'
     }
   }