+++ CHANGES.md
@@ -12,6 +12,7 @@
 
 **Changed:**
 
+- Apps are no longer required to provide their own HTML template. The default template path of `src/index.html` will continue to be used if it exists. If an alternative template is not provided via [`webpack.plugins.html` config](https://github.com/insin/nwb/blob/v0.11.0/docs/Configuration.md#pluginshtml-object), nwb will now fall back to using a basic template.
 - Restored default use of the Babel polyfill in Karma config so tests (and their dependencies) can assume a modern environment.
 
 **Dependencies:**